Move Requests into their own services
